requirements.txt
pandas
python-dotenv
yfinance
numpy
scikit-learn
xgboost
pyyaml
pytest
langgraph
langchain-core
